Move lock decorator to common
