+++ openfl/transport/grpc/aggregator_server.py
@@ -5,27 +5,15 @@
 """AggregatorGRPCServer module."""
 
 import logging
-import threading
 from random import random
 from time import sleep
 
 import grpc
 
 from openfl.protocols import aggregator_pb2, aggregator_pb2_grpc, utils
-from openfl.transport.grpc.common import create_grpc_server, create_header
+from openfl.transport.grpc.common import create_grpc_server, create_header, synchronized
 
 logger = logging.getLogger(__name__)
-
-
-def synchronized(func):
-    """Executes `func` synchronously in a threading lock."""
-    _lock = threading.Lock()
-
-    def wrapper(self, *args, **kwargs):
-        with _lock:
-            return func(self, *args, **kwargs)
-
-    return wrapper
 
 
 class AggregatorGRPCServer(aggregator_pb2_grpc.AggregatorServicer):
